[step2] refactor : 패키지 구분, View 클래스 생성
