+++ src/main/java/ladder/Application.java
@@ -1,0 +1,8 @@
+package ladder;
+
+public class Application {
+
+    public static void main(String[] args) {
+
+    }
+}
