+++ src/main/java/ladder/domain/BarGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface BarGenerator {
