+++ src/main/java/ladder/domain/Player.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class Player {
     private static final int MAXIMUM_NAME_LENGTH = 5;
