+++ src/main/java/ladder/view/InputView.java
@@ -1,0 +1,4 @@
+package ladder.view;
+
+public class InputView {
+}
