+++ src/main/java/ladder/view/OutputView.java
@@ -1,0 +1,4 @@
+package ladder.view;
+
+public class OutputView {
+}
